+++ NEWS.md
@@ -3,6 +3,8 @@
 ## New features
 
 * new `get_dataset_doc()` function for accessing documentation
+* exported `get_dataset_metadata()` function for accessing detailed metadata
+
 ## Improvements
 
 * get_czso_catalogue() is now much faster as it uses the open data catalogue's API instead of donwloading a huge CSV list of all datasets. It is less flexible as it does not allow direct filtering.